hamperbox-starter/requirements.txt
## requirements.txt
django>=4.2
djangorestframework
Pillow

